refactor: :art: Separating thread, comment, and reply for cleaner architecture (include test)

+++ src/Applications/use_case/_test/DeleteCommentUseCase.test.js
@@ -1,24 +1,34 @@
 const ThreadRepository = require('../../../Domains/threads/ThreadRepository');
+const CommentRepository = require('../../../Domains/comments/CommentRepository');
 const DeleteCommentUseCase = require('../DeleteCommentUseCase');
 
 describe('DeleteCommentUseCase', () => {
   it('should orchestrating the delete comment action correctly', async () => {
     const useCasePayload = {
       threadId: 'thread-123',
+      commentId: 'comment-123',
       userId: 'user-123',
     };
 
     const mockThreadRepository = new ThreadRepository();
+    const mockCommentRepository = new CommentRepository();
 
-    mockThreadRepository.deleteComment = jest.fn()
-      .mockImplementation(() => Promise.resolve());
+    mockThreadRepository.verifyThreadById = jest.fn(() => Promise.resolve());
+    mockCommentRepository.verifyCommentOwner = jest.fn(() => Promise.resolve());
+    mockCommentRepository.deleteComment = jest.fn(() => Promise.resolve());
 
     const deleteCommentUseCase = new DeleteCommentUseCase({
       threadRepository: mockThreadRepository,
+      commentRepository: mockCommentRepository,
     });
 
     await deleteCommentUseCase.execute(useCasePayload);
 
-    expect(mockThreadRepository.deleteComment).toBeCalledWith(useCasePayload);
+    expect(mockThreadRepository.verifyThreadById)
+      .toBeCalledWith({ threadId: useCasePayload.threadId });
+    expect(mockCommentRepository.verifyCommentOwner)
+      .toBeCalledWith({ commentId: useCasePayload.commentId, userId: useCasePayload.userId });
+    expect(mockCommentRepository.deleteComment)
+      .toBeCalledWith({ commentId: useCasePayload.commentId, userId: useCasePayload.userId });
   });
 });
